feat(backend): friendRecs refactor
